# api/openapi.yaml
openapi: 3.0.3

info:
  title: Course Task Tracker API
  version: 1.0.0
  description: Basic task tracking API for CS453 midterm


servers:
  - url: http://localhost:3000


paths:

  /health:

    get:
      summary: Check /health
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok


  /api/tasks:

    get:
      summary: get all /api/tasks
      responses:

        "200":
          description: list of tasks 
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"


    post:
      summary: create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTask"

      responses:
        
        "201":
          description: Task created successfully

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "400":
          description: invalid task data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"



  /api/tasks/{id}:

    get:
      summary: get a task by id 

      parameters:
        
        - name: id 
          in: path
          required: true
          description: task id 
          schema:
            type: integer


      responses:

        "200":
          description: task found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"


        "404":
          description: task not found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


    put:
      summary: replace a task


      parameters:
        - name: id 
          in: path
          required: true
          description: task id 
          schema:
            type: integer

      requestBody:
        required: true

        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/NewTask"


      responses:

        "200":
          description: Task replaced successfully

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task" 


        "400":
          description: Invalid task data

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error" 


        "404":
          description: task not found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


    patch:
      summary: Update a task

      parameters:
        - name: id 
          in: path
          required: true
          description: task id 
          schema:
            type: integer


      requestBody:
        required: true

        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"

      responses:

        "200":
          description: Task updated successfully

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task" 


        "400":
          description: Invalid update data

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error" 

        "404":
          description: Task not found

          content:
            application/json:
               schema:
                 $ref: "#/components/schemas/Error"



    delete:
      summary: Delete a task


      parameters: 
        - name: id 
          in: path
          required: true
          description: task id 
          schema:
            type: integer


      responses:

        "204":
          description: task deleted successfully


        "404":
          description: Task not found 
          
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"




components:
  schemas:


    Task: 
      type: object

      required:
        - id 
        - title 
        - course
        - completed 

      properties:

        id: 
          type: integer
          example: 3

        title:
          type: string
          example: Complete CS 453 midterm
        
        course:  
          type: string
          example: CS453

        completed:
          type: boolean
          example: false


    NewTask:

      type: object

      required:
        - title
        - course
        - completed

      properties:


        title: 
          type: string
          example: Comeplete CS453 midterm

        course:
          type: string
          example: CS453

        completed:

          type: boolean
          example: false


    UpdateTask:
      type: object

      properties:
        title:
          type: string
          example: Complete Checkpoint 1

        course:
          type: string
          example: CS453

        completed:
          type: boolean
          example: true


    Error:

      type: object


      properties:


        error:
          
          type: string

          example: cannot find task